chore(refinement): Tighten up api code

+++ JobFlow.Infrastructure.Persistence/Repository.cs
@@ -1,6 +1,8 @@
-﻿using System.Linq.Expressions;
-using JobFlow.Domain;
+﻿using JobFlow.Domain;
+using JobFlow.Domain.Models;
 using Microsoft.EntityFrameworkCore;
+using System.Linq.Expressions;
+using ISoftDeletable = JobFlow.Domain.Models.ISoftDeletable;
 
 namespace JobFlow.Infrastructure.Persistence;
 
@@ -73,22 +75,64 @@ public class Repository<T> : IRepository<T> where T : class
     // 🔹 Delete
     public void Remove(T item)
     {
+        if (item is ISoftDeletable softDeletable)
+        {
+            softDeletable.IsActive = false;
+            softDeletable.DeactivatedAtUtc = DateTime.UtcNow;
+            _dbSet.Update(item);
+            return;
+        }
+
         _dbSet.Remove(item);
     }
 
     public Task RemoveAsync(T item)
     {
+        if (item is ISoftDeletable softDeletable)
+        {
+            softDeletable.IsActive = false;
+            softDeletable.DeactivatedAtUtc = DateTime.UtcNow;
+            _dbSet.Update(item);
+            return Task.CompletedTask;
+        }
+
         _dbSet.Remove(item);
         return Task.CompletedTask;
     }
 
     public void RemoveRange(IEnumerable<T> items)
     {
+        if (typeof(ISoftDeletable).IsAssignableFrom(typeof(T)))
+        {
+            foreach (var item in items)
+            {
+                if (item is not ISoftDeletable softDeletable) continue;
+                softDeletable.IsActive = false;
+                softDeletable.DeactivatedAtUtc = DateTime.UtcNow;
+            }
+
+            _dbSet.UpdateRange(items);
+            return;
+        }
+
         _dbSet.RemoveRange(items);
     }
 
     public Task RemoveRangeAsync(IEnumerable<T> items)
     {
+        if (typeof(ISoftDeletable).IsAssignableFrom(typeof(T)))
+        {
+            foreach (var item in items)
+            {
+                if (item is not ISoftDeletable softDeletable) continue;
+                softDeletable.IsActive = false;
+                softDeletable.DeactivatedAtUtc = DateTime.UtcNow;
+            }
+
+            _dbSet.UpdateRange(items);
+            return Task.CompletedTask;
+        }
+
         _dbSet.RemoveRange(items);
         return Task.CompletedTask;
     }
